# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Implementation Api CV",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/familia": {
            "get": {
                "tags": [
                    "Familia"
                ],
                "description": "Obtain family records data",
                "operationId": "f750083e59bd410f6e12f81948f6d315",
                "responses": {
                    "200": {
                        "description": "Listado de Datos Familiares"
                    }
                }
            },
            "post": {
                "tags": [
                    "Familia"
                ],
                "description": "Store a newly created resource in database",
                "operationId": "a1f8fa86bdffb93b0c7432b399207846",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "description": "Nombre Completo",
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "Género",
                                        "type": "string"
                                    },
                                    "date_of_birth": {
                                        "description": "Fecha de cumpleaños",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "Identificación del Usuario",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Crear un registro familiar"
                    }
                }
            }
        },
        "/api/familia/{id}": {
            "get": {
                "tags": [
                    "Familia"
                ],
                "description": "Get family information by ID",
                "operationId": "a18762f16da76d1d5f77c6962421c389",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener la información familiar por su id"
                    }
                }
            },
            "put": {
                "tags": [
                    "Familia"
                ],
                "description": "Update a record",
                "operationId": "1e55997c009aef3e43396d66f7ec6a0f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "description": "Nombre Completo",
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "Género",
                                        "type": "string"
                                    },
                                    "date_of_birth": {
                                        "description": "Fecha de nacimiento",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "identificación del usuario",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Actualizar un registro familiar"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Familia"
                ],
                "description": "Delete a single type on the ID supplied",
                "operationId": "ab8361c40af1c30e5b9a606c47fbd4e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminar un registro familiar"
                    }
                }
            }
        },
        "/api/salario": {
            "get": {
                "tags": [
                    "Salarios"
                ],
                "description": "Get a list of salaries",
                "operationId": "e1e58db3e7ebe9704c11cdf9db6e15d8",
                "responses": {
                    "200": {
                        "description": "Listado de Salarios"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Salarios"
                ],
                "description": "Store a newly created resource in database",
                "operationId": "043604dc983fd9659a37824d50a9be9c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "salary": {
                                        "description": "Salario",
                                        "type": "number"
                                    },
                                    "bonus": {
                                        "description": "Bonificación 78 89",
                                        "type": "number"
                                    },
                                    "agreed_bonus": {
                                        "description": "Bonificación Acordado Pactada",
                                        "type": "number"
                                    },
                                    "user_id": {
                                        "description": "Identificación del usuario",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Crear un recurso de Salario"
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/salario/{id}": {
            "get": {
                "tags": [
                    "Salarios"
                ],
                "description": "Get a Salary for the id",
                "operationId": "432e35122b52e2dbfb37df396c3bc8c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id for salary",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener un Salario"
                    }
                }
            },
            "put": {
                "tags": [
                    "Salarios"
                ],
                "description": "Update an employee's salary",
                "operationId": "9d8066204d4863137c938169e1a73815",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "salary": {
                                        "description": "Salario",
                                        "type": "integer"
                                    },
                                    "bonus": {
                                        "description": "Bonificación 78 89",
                                        "type": "integer"
                                    },
                                    "agreed_bonus": {
                                        "description": "Bonificación Pactada",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "Identificación del Usuario",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Actualizar un Salario de Empleado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Salarios"
                ],
                "description": "Delete a single salary's on the ID supplied",
                "operationId": "12560aa0de02422c76acbc8d2d8e8364",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of salary to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminar un Salario"
                    }
                }
            }
        },
...
    }
}